Repo and service layer method immplemntation

+++ src/main/java/com/learning/repo/AccountRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface AccountRepo extends JpaRepository<Account,Integer> {
+public interface AccountRepo extends JpaRepository<Account,Long> {
     public List<Account> findByIsApproved(boolean isApproved);
 }
